riecoin 0.8.7 and bitcoin 0.9.1 manually merged

+++ src/wallet.h
@@ -449,8 +449,8 @@ public:
     mapValue_t mapValue;
     std::vector<std::pair<std::string, std::string> > vOrderForm;
     unsigned int fTimeReceivedIsTxTime;
-    unsigned int nTimeReceived;  // time received by this node
-    unsigned int nTimeSmart;
+    int64_t nTimeReceived;  // time received by this node
+    int64_t nTimeSmart;
     char fFromMe;
     std::string strFromAccount;
     int64_t nOrderPos;  // position in ordered transaction list
@@ -543,7 +543,7 @@ public:
 
             ReadOrderPos(pthis->nOrderPos, pthis->mapValue);
 
-            pthis->nTimeSmart = mapValue.count("timesmart") ? (unsigned int)atoi64(pthis->mapValue["timesmart"]) : 0;
+            pthis->nTimeSmart = mapValue.count("timesmart") ? atoi64(pthis->mapValue["timesmart"]) : 0;
         }
 
         pthis->mapValue.erase("fromaccount");
